Create README.md
I tried to write comments to myself to remember for next time, but they may not be super clear. Feel free to ask. It helped me to set the wd from the file pane for some reason, and then when it writes your lists and tells you how many items in them, so can check it got all the files it needed to. Obvs the filenames and stuff are specific to my data for my Poverty Creek sites :)